backend/requirements.txt
Flask
flask-cors
google-generativeai
python-dotenv
youtube-transcript-api
